Stop depending ArrayRecordMacros -> Macros.
